+++ src/app/class/create/address/page.tsx
@@ -1,64 +1,17 @@
-'use client';
-import { FormEvent, useRef, useState } from 'react';
-import { searchAddress } from '@/lib/apis/searchAddress';
-import AddressDescription from './_components/AddressDescription';
-import SearchForm from './_components/SearchForm';
-import SearchResults from './_components/SearchResults';
-import Pagination from '@/components/Pagination/Pagination';
-import { AddressData } from '@/types/address';
+import { headers } from 'next/headers';
+import { redirect } from 'next/navigation';
+import { DOMAIN } from '@/constants/constants';
+import Address from './_components/Address';
 
-const Address = () => {
-  const [addressData, setAddressData] = useState<AddressData | null>(null);
-  const [currentPage, setCurrentPage] = useState(0);
-  const inputAddressRef = useRef('');
+const AddressPage = async () => {
+  const headersList = headers();
+  const referer = headersList.get('referer');
 
-  const addressSearch = async (e: FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
+  if (!referer?.startsWith(`${DOMAIN}/class/create`)) {
+    redirect('/');
+  }
 
-    const formData = new FormData(e.currentTarget);
-    const value = formData.get('inputAddress') as string;
-
-    if (value) {
-      inputAddressRef.current = value;
-      setAddressData(await searchAddress(value, 0));
-      setCurrentPage(0);
-    }
-  };
-
-  const handlePageChange = async ({ selected }: { selected: number }) => {
-    setCurrentPage(selected);
-    setAddressData(await searchAddress(inputAddressRef.current, selected + 1));
-  };
-
-  return (
-    <>
-      <header className="m-auto h-20 w-full max-w-2xl px-11 ">
-        <h1 className="pt-8 text-2xl font-semibold">주소 검색</h1>
-      </header>
-      <hr className="border-t-1 mb-7 w-full border-solid border-sub-color2" />
-
-      <section className="m-auto flex w-full max-w-2xl flex-col px-11">
-        <SearchForm addressSearch={addressSearch} />
-
-        {addressData ? (
-          <SearchResults addressData={addressData} />
-        ) : (
-          <AddressDescription />
-        )}
-
-        {addressData && addressData.results.common.errorCode === '0' && (
-          <Pagination
-            pageCount={Math.ceil(
-              parseInt(addressData.results.common.totalCount) /
-                parseInt(addressData.results.common.countPerPage),
-            )}
-            currentPage={currentPage}
-            onPageChange={handlePageChange}
-          />
-        )}
-      </section>
-    </>
-  );
+  return <Address />;
 };
 
-export default Address;
+export default AddressPage;
